+++ packages/ethereum-storage/package.json
@@ -63,7 +63,6 @@
     "source-map-support": "0.5.19",
     "ts-jest": "26.3.0",
     "ts-node": "10.9.1",
-    "typed-emitter": "1.4.0",
     "typescript": "4.9.5",
     "web3-providers-http": "1.3.6"
   },
