+++ packages/thegraph-data-access/package.json
@@ -46,8 +46,7 @@
     "@requestnetwork/utils": "0.36.0",
     "ethers": "5.5.1",
     "graphql-request": "3.4.0",
-    "tslib": "2.5.0",
-    "typed-emitter": "1.4.0"
+    "tslib": "2.5.0"
   },
   "devDependencies": {
     "@types/jest": "26.0.13",
